aws: create, get, delete, list, invoke

+++ function/function.go
@@ -7,7 +7,7 @@ type Meta struct {
 	Runtime               string           `json:"runtime"`
 	Handler               string           `json:"handler"`
 	Timeout               int32            `json:"timeout"`
-	MemorySize            int32            `json:"memorySize"`
+	MemorySize            int64            `json:"memorySize"`
 	CodeSize              int64            `json:"codeSize"`
 	CodeChecksum          string           `json:"codeChecksum"`
 	EnvironmentVariables  map[string]string `json:"environmentVariables"`
